refactor: modularize routes
